chore: consolidate header and logo components
