--- Logo.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-export const Logo: React.FC = () => {
-  return (
-    <h1 className="text-2xl lg:text-3xl font-bold text-gray-800 dark:text-gray-100 font-serif tracking-tight">
-      jus-tice courts
-    </h1>
-  );
-};
